chore: review Audit files and add tests
