--- apps/ai-dial-admin/src/components/ActivityAudit/Modals/tests.tsx/Details.spec.tsx
@@ -1,13 +1,0 @@
-import { describe, expect, test } from 'vitest';
-import { render, screen } from '@testing-library/react';
-
-import ActivityDetails from '../Details';
-import { PopUpState } from '@/src/types/pop-up';
-
-describe('Activity Details Modal', () => {
-  test('Should render component', async () => {
-    render(<ActivityDetails auditViewId="id" onClose={() => void 0} modalState={PopUpState.Opened} />);
-
-    expect(screen.getByText('ActivityAudit.ActivityDetails')).toBeInTheDocument();
-  });
-});
