+++ apps/ai-dial-admin/src/components/ActivityAudit/constants.ts
@@ -1,62 +1,4 @@
-import { ColDef } from 'ag-grid-community';
-
-import { ActivityAuditResourceType } from '@/src/types/activity-audit';
-import {
-  APPLICATIONS_COLUMNS,
-  KEYS_COLUMNS,
-  RUNNERS_COLUMNS,
-  SIMPLE_ENTITY_COLUMNS,
-  SIMPLE_DESCRIPTION_COLUMNS,
-  MODELS_COLUMNS,
-} from '@/src/constants/grid-columns/grid-columns';
 import { SortDirectionDto } from '@/src/types/request';
-
-export const SYSTEM_ROLLBACK_ID = 'system-rollback';
-export const SYSTEM_ROLLBACK_ENTITIES = [
-  ActivityAuditResourceType.MODEL,
-  ActivityAuditResourceType.APPLICATION,
-  ActivityAuditResourceType.ROUTE,
-  ActivityAuditResourceType.ROLE,
-  ActivityAuditResourceType.KEY,
-  ActivityAuditResourceType.APPLICATION_TYPE_SCHEMA,
-  ActivityAuditResourceType.INTERCEPTOR,
-  ActivityAuditResourceType.ADAPTER,
-];
-
-export const SYSTEM_ROLLBACK_TAB_NAME: Record<ActivityAuditResourceType, string> = {
-  [ActivityAuditResourceType.MODEL]: 'Models',
-  [ActivityAuditResourceType.APPLICATION]: 'Applications',
-  [ActivityAuditResourceType.ADAPTER]: 'Adapters',
-  [ActivityAuditResourceType.INTERCEPTOR]: 'Interceptors',
-  [ActivityAuditResourceType.KEY]: 'Keys',
-  [ActivityAuditResourceType.ROLE]: 'Roles',
-  [ActivityAuditResourceType.APPLICATION_TYPE_SCHEMA]: 'ApplicationRunners',
-  [ActivityAuditResourceType.ROUTE]: 'Routes',
-};
-
-export const getSystemRollbackColumns = (
-  type: ActivityAuditResourceType,
-  t: (stringToTranslate: string) => string,
-): ColDef[] => {
-  switch (type) {
-    case ActivityAuditResourceType.MODEL:
-      return MODELS_COLUMNS(t);
-    case ActivityAuditResourceType.APPLICATION:
-      return APPLICATIONS_COLUMNS(t);
-    case ActivityAuditResourceType.ADAPTER:
-      return SIMPLE_DESCRIPTION_COLUMNS;
-    case ActivityAuditResourceType.INTERCEPTOR:
-    case ActivityAuditResourceType.ROLE:
-    case ActivityAuditResourceType.ROUTE:
-      return SIMPLE_ENTITY_COLUMNS;
-    case ActivityAuditResourceType.KEY:
-      return KEYS_COLUMNS;
-    case ActivityAuditResourceType.APPLICATION_TYPE_SCHEMA:
-      return RUNNERS_COLUMNS;
-    default:
-      return [];
-  }
-};
 
 export const sorts = [
   {
@@ -64,3 +6,101 @@ export const sorts = [
     direction: SortDirectionDto.DESC,
   },
 ];
+
+export enum EntityParameterKeys {
+  TOPICS = 'topics',
+  HASHING_ORDER = 'fieldsHashingOrder',
+  LIMITS = 'limits',
+  PRICING = 'pricing',
+  INTERCEPTORS = 'interceptors',
+  ROLES = 'roles',
+  ROLE_LIMITS = 'roleLimits',
+  DEFAULT_ROLE_LIMIT = 'defaultRoleLimit',
+  UPSTREAMS = 'upstreams',
+  FEATURES = 'features',
+  ENABLED = 'enabled',
+  PROPERTIES = 'properties',
+  PATHS = 'paths',
+  METHODS = 'methods',
+  RESPONSE = 'response',
+  APPLICATIONS = 'applications',
+  ENTITIES = 'entities',
+  DEFS = '$defs',
+  KEYS = 'grantedKeys',
+  PARAMETERS = 'parameters',
+  MODELS = 'models',
+  UNIT = 'unit',
+}
+
+export enum ParameterNamesI18nKey {
+  roleLimits = 'ParameterNames.roleLimits',
+  isPublic = 'ParameterNames.isPublic',
+  defaultRoleLimit = 'ParameterNames.defaultRoleLimit',
+  fieldsHashingOrder = 'ParameterNames.fieldsHashingOrder',
+  endpointDeploymentName = 'ParameterNames.endpointDeploymentName',
+  name = 'ParameterNames.name',
+  author = 'ParameterNames.author',
+  endpoint = 'ParameterNames.endpoint',
+  displayName = 'ParameterNames.displayName',
+  displayVersion = 'ParameterNames.displayVersion',
+  adapter = 'ParameterNames.adapter',
+  iconUrl = 'ParameterNames.iconUrl',
+  description = 'ParameterNames.description',
+  forwardAuthToken = 'ParameterNames.forwardAuthToken',
+  features = 'ParameterNames.features',
+  inputAttachmentTypes = 'ParameterNames.inputAttachmentTypes',
+  maxInputAttachments = 'ParameterNames.maxInputAttachments',
+  defaults = 'ParameterNames.defaults',
+  topics = 'ParameterNames.topics',
+  maxRetryAttempts = 'ParameterNames.maxRetryAttempts',
+  type = 'ParameterNames.type',
+  tokenizerModel = 'ParameterNames.tokenizerModel',
+  limits = 'ParameterNames.limits',
+  maxTotalTokens = 'ParameterNames.maxTotalTokens',
+  pricing = 'ParameterNames.pricing',
+  unit = 'ParameterNames.unit',
+  prompt = 'ParameterNames.prompt',
+  completion = 'ParameterNames.completion',
+  upstreams = 'ParameterNames.upstreams',
+  key = 'ParameterNames.key',
+  weight = 'ParameterNames.weight',
+  tier = 'ParameterNames.tier',
+  overrideName = 'ParameterNames.overrideName',
+  properties = 'ParameterNames.properties',
+  enabled = 'ParameterNames.enabled',
+  minute = 'ParameterNames.minute',
+  day = 'ParameterNames.day',
+  week = 'ParameterNames.week',
+  month = 'ParameterNames.month',
+  interceptors = 'ParameterNames.interceptors',
+  roles = 'ParameterNames.roles',
+  paths = 'ParameterNames.paths',
+  methods = 'ParameterNames.methods',
+  status = 'ParameterNames.status',
+  body = 'ParameterNames.body',
+  $defs = 'ParameterNames.$defs',
+  $id = 'ParameterNames.$id',
+  $schema = 'ParameterNames.$schema',
+  applications = 'ParameterNames.applications',
+  entities = 'ParameterNames.entities',
+  'dial:applicationTypeCompletionEndpoint' = 'ParameterNames.applicationTypeCompletionEndpoint',
+  'dial:applicationTypeDisplayName' = 'ParameterNames.applicationTypeDisplayName',
+  'dial:applicationTypeEditorUrl' = 'ParameterNames.applicationTypeEditorUrl',
+  'dial:applicationTypeViewerUrl' = 'ParameterNames.applicationTypeViewerUrl',
+  'dial:applicationTypeConfigurationEndpoint' = 'ParameterNames.applicationTypeConfigurationEndpoint',
+  'dial:applicationTypeRateEndpoint' = 'ParameterNames.applicationTypeRateEndpoint',
+  'dial:applicationTypeTokenizeEndpoint' = 'ParameterNames.applicationTypeTokenizeEndpoint',
+  'dial:applicationTypeTruncatePromptEndpoint' = 'ParameterNames.applicationTypeTruncatePromptEndpoint',
+  'dial:appendApplicationPropertiesHeader' = 'ParameterNames.appendApplicationPropertiesHeader',
+  grantedKeys = 'ParameterNames.grantedKeys',
+  createdAt = 'ParameterNames.createdAt',
+  updatedAt = 'ParameterNames.updatedAt',
+  expiresAt = 'ParameterNames.expiresAt',
+  keyGeneratedAt = 'ParameterNames.keyGeneratedAt',
+  project = 'ParameterNames.project',
+  projectContactPoint = 'ParameterNames.projectContactPoint',
+  secured = 'ParameterNames.secured',
+  parameters = 'ParameterNames.parameters',
+  extraData = 'ParameterNames.extraData',
+  models = 'ParameterNames.models',
+}
